requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
python-decouple==3.8
Pillow==9.5.0
drf-yasg==1.21.7
